Refactor page evaluation stages and outcomes
Split pipeline stages and test fixtures into focused files, and
represent terminal outcomes with typed classifications.
