+++ Tests/ClawEvaluationTests/Security/EvaluationFilesystemSecurityTests.swift
@@ -1,0 +1,3 @@
+import Testing
+
+@Suite struct EvaluationFilesystemSecurityTests {}
